playercontrols seperated from player for reuse
